+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.59.0 (2026-04-13)
+
+### Feat
+
+- **traceloop-sdk**: Add guardrails (#3649)
+
+## v0.58.1 (2026-04-12)
+
+### Fix
+
+- **evaluators**: update route api (#3977)
+
 ## v0.58.0 (2026-04-09)
 
 ### Feat